agent: move structs into consul/structs pkg
 * CheckDefinition
 * ServiceDefinition
 * CheckType
