+++ agent/consul/structs/check_defintion.go
@@ -1,50 +1,11 @@
-package agent
+package structs
 
 import (
 	"time"
 
-	"github.com/hashicorp/consul/agent/consul/structs"
 	"github.com/hashicorp/consul/api"
 	"github.com/hashicorp/consul/types"
 )
-
-// ServiceDefinition is used to JSON decode the Service definitions
-type ServiceDefinition struct {
-	ID                string
-	Name              string
-	Tags              []string
-	Address           string
-	Port              int
-	Check             CheckType
-	Checks            CheckTypes
-	Token             string
-	EnableTagOverride bool
-}
-
-func (s *ServiceDefinition) NodeService() *structs.NodeService {
-	ns := &structs.NodeService{
-		ID:                s.ID,
-		Service:           s.Name,
-		Tags:              s.Tags,
-		Address:           s.Address,
-		Port:              s.Port,
-		EnableTagOverride: s.EnableTagOverride,
-	}
-	if ns.ID == "" && ns.Service != "" {
-		ns.ID = ns.Service
-	}
-	return ns
-}
-
-func (s *ServiceDefinition) CheckTypes() (checks CheckTypes) {
-	s.Checks = append(s.Checks, &s.Check)
-	for _, check := range s.Checks {
-		if check.Valid() {
-			checks = append(checks, check)
-		}
-	}
-	return
-}
 
 // CheckDefinition is used to JSON decode the Check definitions
 type CheckDefinition struct {
@@ -74,8 +35,8 @@ type CheckDefinition struct {
 	DeregisterCriticalServiceAfter time.Duration
 }
 
-func (c *CheckDefinition) HealthCheck(node string) *structs.HealthCheck {
-	health := &structs.HealthCheck{
+func (c *CheckDefinition) HealthCheck(node string) *HealthCheck {
+	health := &HealthCheck{
 		Node:      node,
 		CheckID:   c.ID,
 		Name:      c.Name,
@@ -109,11 +70,4 @@ func (c *CheckDefinition) CheckType() *CheckType {
 		Status: c.Status,
 		Notes:  c.Notes,
 	}
-}
-
-// persistedService is used to wrap a service definition and bundle it
-// with an ACL token so we can restore both at a later agent start.
-type persistedService struct {
-	Token   string
-	Service *structs.NodeService
 }
